Modify users and transactions operations
